// a3front/src/modules/negotiations/products/general/enums/service-type.enum.ts
export enum ServiceTypeEnum {
  FOOD = 1, // Alimentación
  ACTIVITIES = 2, // Actividades
  EXCURSIONS = 3, // Excursiones
  TRANSFERS = 4, // Traslados
  ASSISTANCE = 5, // Asistencia
  MULTIDAYS = 6, // Multidias
  ENTRIES = 7, // Entradas
  GUIDE = 8, // Guia
  CRUISES = 9, // Cruceros
  TRANSPORT = 10, // Transporte
  BUS_TICKET = 11, // Ticket Bus
  AIR_TICKET = 12, // Ticket Aereo
  TRAIN_TICKET = 13, // Ticket Tren
  OTHERS = 14, // Otros
}